requirements.txt
opencv-python==4.7.0.72
dlib==19.24.0
mediapipe==0.10.13
tensorflow-cpu==2.18.0
numpy>=1.26,<2.0
matplotlib
pandas
